# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Renting.ru API
  version: 1.0.0
  description: |
    API-first rental platform. All responses use the standard envelope:
    `{ success, data, error: { code, message }, meta: { pagination } }`.
    Stable machine-readable `error.code` values are defined in `@renting/shared` `ErrorCode`.
  contact:
    name: Renting.ru
    email: api@renting.ru
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
  - url: https://api.renting.ru/api/v1
    description: Production
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
...
paths:
...
  /bookings/{id}:
    get:
      tags: [Bookings]
      summary: Get booking detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Full booking with listing, driver, payment, invoice
...
  /pages/{slug}:
    get:
      tags: [Content]
      summary: Get CMS content page
      security: []
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string, example: about }
      responses:
        '200':
          description: Page content
...
  /admin/bookings:
    get:
      tags: [Admin — Bookings]
      summary: List all bookings
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: perPage
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Paginated bookings
...
  /admin/coupons:
    get:
      tags: [Admin — Coupons]
      summary: List coupons
      responses:
        '200':
          description: Coupons

    post:
      tags: [Admin — Coupons]
      summary: Create a coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, discountType, discountValue]
              properties:
                code: { type: string }
                discountType: { type: string, enum: [percent, fixed] }
                discountValue: { type: number }
                maxDiscount: { type: number }
                usageLimit: { type: integer }
                validFrom: { type: string, format: date-time }
                validUntil: { type: string, format: date-time }
                isActive: { type: boolean }
      responses:
        '201':
          description: Coupon created

  # ─── Admin — Reports ────────────────────────────────────
...